Updates mesh reader. Adds meshbuilder tests.

+++ src/Shadowbane.Models/Index.cs
@@ -2,15 +2,15 @@
 {
     public class Index
     {
-        public Index(int position, int textureCoordinate, int normal)
+        public Index(ushort position, ushort textureCoordinate, ushort normal)
         {
             this.Position = position;
             this.TextureCoordinate = textureCoordinate;
             this.Normal = normal;
         }
 
-        public int Position { get; set; }
-        public int Normal { get; set; }
-        public int TextureCoordinate { get; set; }
+        public ushort Position { get; set; }
+        public ushort Normal { get; set; }
+        public ushort TextureCoordinate { get; set; }
     }
 }
